requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
passlib[bcrypt]
python-jose
